Move golden utils to standard library
